Add load_wasm_client helper
