+++ relayer/crates/starknet-integration-tests/src/utils/mod.rs
@@ -1,3 +1,5 @@
 mod init_bootstrap;
+mod wasm_client;
 
 pub use init_bootstrap::*;
+pub use wasm_client::*;
